Stack 2 - Dagster + DuckDB + S3 + Terraform (#1)
Creates the first version for the following stack:

Dagster
DuckDB
S3
Terraform
Localstack

This is a simple example following the article Build a poor man’s data lake from scratch with DuckDB. It is a good entry point for learning more about dagster. I added a docker-compose file for running localstack which adds the option to create a "persistent" s3 bucket if it is needed. I also added terraform in order to create all resources that are needed for the AWS implementation.